AP-1149 Extract monthly equivalency calculation out to a concern
