+++ app/models/other_income_source.rb
@@ -1,4 +1,6 @@
 class OtherIncomeSource < ApplicationRecord
+  include MonthlyEquivalentCalculator
+
   belongs_to :gross_income_summary
   has_many :other_income_payments
 
@@ -7,29 +9,7 @@ class OtherIncomeSource < ApplicationRecord
   delegate :assessment, to: :gross_income_summary
 
   def calculate_monthly_income!
-    assessment.assessment_errors.create!(record_id: id, record_type: self.class, error_message: converter.error_message) if converter.error?
-
-    update!(monthly_income: converter.monthly_amount)
-    converter.monthly_amount
-  end
-
-  private
-
-  def dates_and_amounts
-    Utilities::PaymentPeriodDataExtractor.call(collection: other_income_payments,
-                                               date_method: :payment_date,
-                                               amount_method: :amount)
-  end
-
-  def frequency
-    Utilities::PaymentPeriodAnalyser.new(dates_and_amounts).period_pattern
-  end
-
-  def converter
-    @converter ||= Calculators::UnearnedIncomeMonthlyConvertor.new(frequency, payment_amounts)
-  end
-
-  def payment_amounts
-    other_income_payments.map(&:amount)
+    calculate_monthly_equivalent!(target_field: :monthly_income,
+                                  collection: other_income_payments)
   end
 end
